Refactor app.helpers
