+++ app/wiki/routes.py
@@ -1,5 +1,6 @@
 from app import db
-from app.helpers import page_title, urlfriendly, deny_access, moderator_required
+from app.decorators import moderator_required
+from app.helpers import page_title, urlfriendly, deny_access
 from app.map.helpers import get_nodes_by_wiki_id
 from app.wiki import bp
 from app.wiki.forms import WikiEntryForm, WikiSearchForm, WikiMoveCategoryForm
